Fix proximity in low power proximity
Android considers the values 1->5 to be a released proximity event and 0 to be a triggered event. Meanwhile the driver reports 1 -> 3 when proximity is triggered when the screen is powered on and 0 when proximity is released. When in low power proximity mode, the driver reports 4 for a triggered proximity event and 5 for a released proximity event. This patch solves this by reporting 0 when the value would be reported as 1->4 and 5 when the value is either 0 or 5.